Move useLayoutState composable to shared
