+++ app/src/interfaces/list-o2m-tree-view/options.vue
@@ -22,9 +22,8 @@
 
 <script lang="ts">
 import { useI18n } from 'vue-i18n';
-import { Field } from '@/types';
+import { Field, Relation } from '@/types';
 import { defineComponent, PropType, computed } from 'vue';
-import { Relation } from '@/types/relations';
 export default defineComponent({
 	emits: ['input'],
 	props: {
