+++ app/src/interfaces/select-dropdown-m2o/options.vue
@@ -12,9 +12,8 @@
 
 <script lang="ts">
 import { useI18n } from 'vue-i18n';
-import { Field } from '@/types';
+import { Field, Relation } from '@/types';
 import { defineComponent, PropType, computed } from 'vue';
-import { Relation } from '@/types/relations';
 
 export default defineComponent({
 	emits: ['input'],
