+++ app/src/layouts/cards/sidebar.vue
@@ -11,7 +11,7 @@
 <script lang="ts">
 import { defineComponent, toRefs } from 'vue';
 
-import { useLayoutState } from '@/composables/use-layout';
+import { useLayoutState } from '@directus/shared/composables';
 
 export default defineComponent({
 	setup() {
