+++ app/src/layouts/tabular/options.vue
@@ -54,7 +54,7 @@ import { useI18n } from 'vue-i18n';
 import { defineComponent, toRefs } from 'vue';
 
 import Draggable from 'vuedraggable';
-import { useLayoutState } from '@/composables/use-layout';
+import { useLayoutState } from '@directus/shared/composables';
 
 export default defineComponent({
 	components: { Draggable },
