+++ app/src/modules/settings/routes/data-model/field-detail/store.ts
@@ -10,7 +10,8 @@ import { DisplayConfig } from '@/displays/types';
 import { getInterfaces } from '@/interfaces';
 import { InterfaceConfig } from '@/interfaces/types';
 import { useCollectionsStore, useFieldsStore, useRelationsStore } from '@/stores/';
-import { Collection, Field, localTypes, Relation, Item } from '@/types';
+import { Collection, Field, localTypes, Relation } from '@/types';
+import { Item } from '@directus/shared/types';
 import { clone, throttle } from 'lodash';
 import { computed, ComputedRef, nextTick, reactive, watch, WatchStopHandle } from 'vue';
