+++ app/src/types/index.ts
@@ -1,9 +1,7 @@
 export * from './collections';
 export * from './error';
 export * from './fields';
-export * from './items';
 export * from './notifications';
 export * from './permissions';
-export * from './presets';
 export * from './relations';
 export * from './users';
