+++ app/src/views/private/components/filter-sidebar-detail/filter-sidebar-detail.vue
@@ -43,7 +43,8 @@
 <script lang="ts">
 import { useI18n } from 'vue-i18n';
 import { defineComponent, PropType, computed, ref, watch, toRefs } from 'vue';
-import { Field, Filter } from '@/types';
+import { Field } from '@/types';
+import { Filter } from '@directus/shared/types';
 import { useFieldsStore } from '@/stores';
 import FieldFilter from './field-filter.vue';
 import { nanoid } from 'nanoid';
