+++ packages/shared/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './extensions';
+export * from './symbols';
