+++ packages/shared/src/constants/symbols.ts
@@ -1,0 +1,1 @@
+export const LAYOUT_SYMBOL = Symbol();
