+++ packages/shared/src/types/index.ts
@@ -1,2 +1,5 @@
 export * from './extensions';
+export * from './items';
+export * from './layouts';
 export * from './misc';
+export * from './presets';
